Move guard to guards folder
