+++ src/app/app.module.ts
@@ -24,7 +24,7 @@ import { LineChartComponent } from './components/line-chart/line-chart.component
 import { EventsAnalyticsComponent } from './components/events-analytics/events-analytics.component';
 import { CircularSpinnerComponent } from './components/circular-spinner/circular-spinner.component';
 import { LoginComponent } from './components/login/login.component';
-import {AuthGuard} from './services/authentication/authentication.guard';
+import {AuthGuard} from './guards/authentication/authentication.guard';
 import {AuthenticationService} from './services/authentication/authentication.service';
 
 @NgModule({
